refactor: reorganize src/components/ into domain folders (Phase 5a)
Pure restructuring. 27 loose components moved from src/components/
root into 6 domain folders. File contents are byte-equivalent before
and after — only `git mv` + import path updates.

Created:
  src/components/layout/      7 files (top-level chrome — AppLayout,
                                AppSidebar, NavLink, MobileNav,
                                NotificationBell, SessionTimeout,
                                OnboardingModal)
  src/components/shared/      4 files (generic primitives crossing
                                domains — Avatar, DocumentStatusBadge,
                                DatePicker, TimePicker)
  src/components/coordinator/ 4 files (coordinator-only —
                                CoordinatorHoursConfirmation,
                                VolunteerActivityTab,
                                DepartmentVolunteersTab,
                                RestrictVolunteerModal)

Extended:
  src/components/admin/        6 → 10 (AdminQRCheckIn,
                                DepartmentCoordinatorManager,
                                VolunteerLeaderboard,
                                VolunteerReliabilityBadge)
  src/components/volunteer/    9 → 16 (BookedSlotsDisplay,
                                VolunteerImpactCharts,
                                VolunteerHoursLetter,
                                SlotSelectionDialog,
                                RecommendedShifts, InviteFriendModal,
                                WeatherForecast)
  src/components/shifts/       3 → 4 (InviteVolunteerModal)

Unchanged:
  src/components/ui/           49 shadcn (per rules — don't move)
  src/components/messaging/    5 files (already organized)
  src/components/settings/     10 files (PR #127)
  src/components/checkin/      5 files (PR #131)

No auth/ folder created — no shared auth components exist; the auth
flow lives in pages, and the only auth-flavored sub-component
(LoginForm) is already in checkin/ where it belongs.

Five orphaned components left at src/components/ root with zero
importers anywhere in src/:
  - CoverageAlert.tsx (fixed in PR #120 but unused)
  - OnboardingChecklist.tsx (was imported by VolunteerDashboard
    pre-PR #125, dropped during refactor)
  - ShiftSlotTimeline.tsx (fixed in PR #120 but unused)
  - VolunteerSettings.tsx
  - WaitlistStatus.tsx

Per the rules, these were NOT deleted in this PR — they need
investigation. Tracked in #132. Two were just fixed in PR #120, so
they may be features that got partially reverted or planned but never
wired up; don't assume safe to delete without checking history.

Verification: tsc --build clean, eslint --max-warnings=0 clean,
103/103 vitest passes after every batch.

Phase 5a complete. Phase 5b (form/auth tests) is next.
