+++ src/App.tsx
@@ -4,9 +4,9 @@ import { Toaster as Sonner } from "@/components/ui/sonner";
 import { Toaster } from "@/components/ui/toaster";
 import { TooltipProvider } from "@/components/ui/tooltip";
 import { AuthProvider, useAuth } from "@/contexts/AuthContext";
-import { AppLayout } from "@/components/AppLayout";
-import { SessionTimeout } from "@/components/SessionTimeout";
-import OnboardingModal from "@/components/OnboardingModal";
+import { AppLayout } from "@/components/layout/AppLayout";
+import { SessionTimeout } from "@/components/layout/SessionTimeout";
+import OnboardingModal from "@/components/layout/OnboardingModal";
 import Auth from "./pages/Auth";
 import ForgotPassword from "./pages/ForgotPassword";
 import ResetPassword from "./pages/ResetPassword";
