+++ src/components/admin/AdminQRCheckIn.tsx
