+++ src/components/coordinator/DepartmentVolunteersTab.tsx
@@ -5,7 +5,7 @@ import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
 import { ShieldCheck } from "lucide-react";
 import { useToast } from "@/hooks/use-toast";
-import { RestrictVolunteerModal } from "@/components/RestrictVolunteerModal";
+import { RestrictVolunteerModal } from "@/components/coordinator/RestrictVolunteerModal";
 
 interface Props {
   departmentIds: string[];
