+++ src/components/layout/AppSidebar.tsx
@@ -1,5 +1,5 @@
 import { useAuth } from "@/contexts/AuthContext";
-import { NavLink } from "@/components/NavLink";
+import { NavLink } from "@/components/layout/NavLink";
 import { useLocation } from "react-router-dom";
 import {
   Sidebar, SidebarContent, SidebarGroup, SidebarGroupContent, SidebarGroupLabel,
