+++ src/components/volunteer/SlotSelectionDialog.tsx
@@ -9,7 +9,7 @@ import { Calendar, Users } from "lucide-react";
 import { format } from "date-fns";
 import { useToast } from "@/hooks/use-toast";
 import { formatSlotRange, slotHours } from "@/lib/slot-utils";
-import { WeatherForecast } from "@/components/WeatherForecast";
+import { WeatherForecast } from "@/components/volunteer/WeatherForecast";
 import { sendEmail } from "@/lib/email-utils";
 import { getShiftTimes, timeLabel } from "@/lib/calendar-utils";
