+++ src/pages/DocumentCompliance.tsx
@@ -7,7 +7,7 @@ import { Input } from "@/components/ui/input";
 import { Textarea } from "@/components/ui/textarea";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Badge } from "@/components/ui/badge";
-import { DocumentStatusBadge } from "@/components/DocumentStatusBadge";
+import { DocumentStatusBadge } from "@/components/shared/DocumentStatusBadge";
 import {
   Select, SelectContent, SelectItem, SelectTrigger, SelectValue,
 } from "@/components/ui/select";
